requirements.txt
fastapi==0.115.4
gunicorn==23.0.0
python-socketio==5.11.4
uvicorn==0.32.0
